requirements.txt
requests
flask
irc
python-dotenv
websockets